+++ src/prepareNeuralLM.cpp
@@ -78,7 +78,7 @@ void writeNgrams(const string &input_filename,
      bool ngramize, 
 		 const string &filename,
      int train_data_size,
-		 vector<float> &sent_weights,
+		 vector<double> &sent_weights,
 		 const string &sent_weights_filename)
 {
     ofstream file(filename.c_str());
@@ -532,7 +532,7 @@ int main(int argc, char *argv[])
         }
     }
     //READING SENTENCE WEIGHTS IF THERE ARE ANY
-    vector<float> sent_weights;
+    vector<double> sent_weights;
     if (sent_weights_text != "") {
       cerr<<"Reading sentence weights from "<<sent_weights_text<<endl;
       ifstream sent_weights_file(sent_weights_text.c_str());
